Multiple improvements

+++ MUD2/packages/contracts/src/systems/TokenSystem.sol
@@ -9,7 +9,7 @@ import { Crawl } from "../utils/Crawl.sol";
 contract TokenSystem is System {
   
   // Bridge setters
-  function setTokenIdToCoord(uint256 tokenId, uint256 coord) public {
+  function setTokenIdToCoord(uint32 tokenId, uint256 coord) public {
     Token.set(tokenId, coord);
   }
